Introduced binding/ dir, splited structs and attach_function
